move theme to file and add btn disabled style
